+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud/compare/v3.12.1...HEAD)
 ------------------------------------------------------------------------
+### Added
+- Adds `FixAll` plugin
+
 ### Change
 - Updates `Fix.__call__` for more flexible error handling